backend/requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pytesseract
Pillow

firebase-admin
chromadb
sentence-transformers
pydantic
pydantic-settings
google-genai
tavily-python
openai-whisper
